+++ packages/klingai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/klingai
 
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies [58bc42d]
+  - @ai-sdk/provider-utils@4.0.16
+
 ## 3.0.4
 
 ### Patch Changes